improved big-endian milo support, hopefully I didn't screw this up <- yeah I did BUT I CAN FIX IT
git-svn-id: svn://japaneatahand.com/rawksd/trunk@150 336e800a-23d8-41f8-b78d-e9b071a98a33

+++ consolehaxx/ConsoleHaxx.Harmonix/FaceFX.cs
@@ -115,7 +115,7 @@ namespace ConsoleHaxx.Harmonix
 			foreach (string str in Strings2)
 				WriteString(str, writer);
 				
-			writer.Write((byte)0x14);
+			writer.Write((int)0x14);
 
 			foreach (int i in Ints2)
 				writer.Write(i);
